pyproject.toml
[project]
name = "supply-inventory"
version = "0.1.0"
description = "ระบบตรวจสอบคลังอุปกรณ์"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1",
    "flask-sqlalchemy>=3.1",
    "flask-login>=0.6",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "python-barcode[images]>=0.15",
    "fpdf2>=2.8",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.5",
]
